Fix cyclical import in ts node
