+++ src/shadow/npm/arborist/lib/dep-valid.mts
@@ -6,9 +6,22 @@ import type { SafeNode } from './node.mts'
 
 const require = createRequire(import.meta.url)
 
-export const depValid: (
+type DepValidFn = (
   child: SafeNode,
   requested: string,
   accept: string | undefined,
   requester: SafeNode
-) => boolean = require(getArboristDepValidPath())
+) => boolean
+
+let _depValid: DepValidFn | undefined
+export function depValid(
+  child: SafeNode,
+  requested: string,
+  accept: string | undefined,
+  requester: SafeNode
+) {
+  if (_depValid === undefined) {
+    _depValid = require(getArboristDepValidPath()) as DepValidFn
+  }
+  return _depValid(child, requested, accept, requester)
+}
